Modifications

+++ API/InventoryManagerAPI/DTO/ProductDTO.cs
@@ -9,12 +9,12 @@ namespace InventoryManagerAPI.DTO
 
         public string description { get; set; }
 
-        public long price { get; set; }
+        public decimal price { get; set; }
 
         public int qty { get; set; }
 
         public QuantityUnit unit { get; set; }
 
-        public long total { get; set; }
+        public decimal total { get; set; }
     }
 }
